Remove redundant wifi test script, incorporating logic and restructuring app functionality
In a major restructuring, the redundant wifi test script, '01_wifi_test_get_uppdata.py', was removed. The wifi communication's logic was integrated into 'main.py', where it forms an integral part of data collection process for gas and temperature detection.

New files were added: 'fit_smoke.py' and 'fit_fire.py' under 'me/src' directory. These files contain functions to perform smoke and fire detection respectively, using the captured images. Another file 'fit_smoke.py' was added under 'playground_thaweewr/pitak' directory. It seems to be a simpler version of the one in the source directory and could be a test script for the functionality.

The 'main.py' was also initialized. It was enhanced to trigger smoke and fire detection and subsequently trigger the model for prediction.

A new directory 'models' was added in the project folder. Its purpose is not clear in the current diff but it might play integral role down the line.

Some minor changes were performed on '01_wifi_test_get.py' file. Functionality was included to create folder in root